refactor: change folder for types, buttons fixes
