+++ travel-app/src/components/Authorization/Common/Types.tsx
